IALERT-3575: Move certificate keystore files to api-certificates.
