+++ api-certificates/build.gradle
@@ -1,0 +1,19 @@
+ext.moduleName = 'com.synopsys.integration.alert.api-certificates'
+
+dependencies {
+    implementation platform(project(':alert-platform'))
+
+    implementation project(':alert-common')
+    implementation project(':api-common-model')
+    implementation project(':alert-database')
+
+    implementation 'org.springframework.boot:spring-boot'
+
+    implementation 'com.synopsys.integration:blackduck-common'
+
+    testImplementation 'org.springframework.data:spring-data-commons'
+    testImplementation project(':test-common-channel')
+    testImplementation project(':provider-blackduck')
+    testImplementation 'com.synopsys.integration:blackduck-common'
+
+}
