+++ component/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     implementation project(':alert-database')
     implementation project(':alert-database-job')
     implementation project(':api-authentication')
+    implementation project(':api-certificates')
     implementation project(':api-common-model')
     implementation project(':api-descriptor')
     implementation project(':api-distribution')
