+++ settings.gradle
@@ -7,6 +7,7 @@ include 'api-authentication'
 include 'api-channel'
 include 'api-channel-issue-tracker'
 include 'api-channel-jira'
+include 'api-certificates'
 include 'api-common-model'
 include 'api-descriptor'
 include 'api-distribution'
