+++ src/main/java/com/synopsys/integration/alert/startup/component/AlertClientCertificateManagerStartupComponent.java
@@ -8,10 +8,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
+import com.synopsys.integration.alert.api.certificates.AlertClientCertificateManager;
 import com.synopsys.integration.alert.api.common.model.exception.AlertException;
 import com.synopsys.integration.alert.common.persistence.model.ClientCertificateKeyModel;
 import com.synopsys.integration.alert.common.persistence.model.ClientCertificateModel;
-import com.synopsys.integration.alert.component.certificates.AlertClientCertificateManager;
 import com.synopsys.integration.alert.database.api.DefaultClientCertificateAccessor;
 import com.synopsys.integration.alert.database.api.DefaultClientCertificateKeyAccessor;
