+++ src/test/java/com/synopsys/integration/alert/component/certificates/AlertSSLContextManagerTestIT.java
@@ -15,6 +15,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.TestPropertySource;
 
+import com.synopsys.integration.alert.api.certificates.AlertClientCertificateManager;
+import com.synopsys.integration.alert.api.certificates.AlertSSLContextManager;
+import com.synopsys.integration.alert.api.certificates.AlertTrustStoreManager;
 import com.synopsys.integration.alert.api.common.model.exception.AlertException;
 import com.synopsys.integration.alert.common.AlertProperties;
 import com.synopsys.integration.alert.common.persistence.model.ClientCertificateKeyModel;
